Чинит провал CI: gomobile Long вместо Int в Kotlin, два гоночных теста, немой build.sh
Реальные причины красных прогонов в Actions.

1. Android не собирался вообще: gomobile отдаёт поля Go-типа int в Kotlin
   как Long, а не Int (getSshPort()/getPoolSize() из ParsedConfig). Строки
   "p.sshPort = if (sshPort > 0) sshPort else 22" пытались присвоить Long
   в поле Int — компилятор Kotlin отказывался с "Assignment type
   mismatch". Локально это не ловилось: у меня нет Android SDK, только
   go build/go test на Go-части.

2. build.sh перенаправлял вывод "go test ./..." в /dev/null, поэтому при
   падении в CI не было видно даже, какой тест провалился — только голый
   exit code 1. Больше не прячем.

3. Два теста гоняли по угаданному времени, а не по реальной синхронизации,
   и оба стабильно проходили в изоляции, но иногда падали под нагрузкой
   всего пакета разом (ровно как на раннере CI):
   - TestKickReconnectsWithoutWaitingCheckInterval проверял, что слот сразу
     после Kick() ещё nil — но обнуляет его фоновая горутина, и под
     нагрузкой она иногда успевала переподключиться раньше следующей строки
     теста. Убрал гонку из самой проверки: она проверяла деталь реализации,
     а не поведение (быстрое переподключение и так проверяется дальше).
   - TestWatchFailoverSwitchesOnSustainedReconnecting спал 50мс, «чтобы
     горутина успела подписаться на шину», и слал событие один раз — если
     горутина не успевала подписаться за эти 50мс, событие терялось
     навсегда (обычная семантика pub/sub) и тест либо зависал, либо не
     находил перехода на запасной сервер. Теперь событие шлётся раз в 10мс
     до самого завершения — reconnectingSince внутри watchFailover
     выставляется по первому дошедшему событию, каким бы оно ни было по
     счёту, так что лишние повторы ничего не портят.

Прогнано: go test ./... -race шесть раз подряд без единого падения, plus
go build/GOOS=windows go build, android/core отдельно.

+++ android/app/src/main/kotlin/io/github/vitazgio/sshtunnel/MainActivity.kt
@@ -634,9 +634,11 @@ class MainActivity : AppCompatActivity() {
         val name = parsed.getName()
         val flag = parsed.getFlag()
         val host = parsed.getHost()
-        val sshPort = parsed.getSshPort()
+        // gomobile отдаёт поля Go-типа int как Long — приводим сразу здесь,
+        // чтобы ниже не путаться с типами при сравнении и присваивании в Int.
+        val sshPort = parsed.getSshPort().toInt()
         val user = parsed.getUser()
-        val poolSize = parsed.getPoolSize()
+        val poolSize = parsed.getPoolSize().toInt()
         val filterMode = parsed.getFilterMode()
         val filterApps = parsed.getFilterApps()
         val directHosts = parsed.getDirectHosts()
